requirements.txt
apscheduler~=3.6.3
requests~=2.25.1
python-telegram-bot
lxml
beautifulsoup4~=4.9.3
grab
psycopg2
sqlalchemy~=1.4.9
pytube
moviepy
pillow
python-dotenv~=0.17.0
